feat!: specify external hosts for mapping
It's really hard to know which interface on a machine is externally
routable so require the user to specify it in the best-case scenario.

Otherwise use `mapAll` to try every available interface until we find
some that work.

Also splits UPnP IPv4 and IPv6 into separate implementations for
better support.

BREAKING CHANGE: map now requires the internal host to map ports for